% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cons_agg.R
\name{create_endpoint_agg}
\alias{create_endpoint_agg}
\title{Create an endpoint for fetching aggregated consumption data}
\usage{
create_endpoint_agg(
  mpan,
  serial,
  period_from = "2024-01-01T00:00:00",
  group_by = "day"
)
}
\arguments{
\item{mpan}{A string representing the MPAN (Meter Point Administration Number).}

\item{serial}{A string representing the meter serial number.}

\item{period_from}{A string representing the start date for the consumption data (default is "2024-01-01T00:00:00").}

\item{group_by}{A string representing the aggregation period (default is "day").
@return A string representing the complete URL endpoint for fetching aggregated consumption data.}
}
\description{
Create an endpoint for fetching aggregated consumption data
}
\details{
The function constructs a URL that can be used to fetch aggregated consumption data for a specific electricity meter point.
The URL includes parameters for pagination, a specific time period, and the aggregation period.
}
\examples{
create_endpoint_agg("1234567890123", "ABC123456789", "2024-01-01T00:00:00", "day")
}
