% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ev.R
\name{evd_process}
\alias{evd_process}
\title{Process EV data}
\usage{
evd_process(df)
}
\arguments{
\item{df}{Data frame containing EV data}
}
\value{
A data frame with processed EV data
}
\description{
Process EV data
}
\examples{
ev_data <- read.csv("ev_data.csv")
evd_process(ev_data)
}
