% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consumption_details.R
\name{get_cons}
\alias{get_cons}
\title{Get Consumption Details
This function retrieves consumption details from the Octopus Energy API.}
\usage{
get_cons(url)
}
\arguments{
\item{url}{The URL to the Octopus Energy API endpoint for consumption details.}
}
\value{
A tibble containing the consumption details.
}
\description{
Get Consumption Details
This function retrieves consumption details from the Octopus Energy API.
}
\examples{
url <- "https://api.octopus.energy/v1/products/your_product_id/consumption/"
get_cons(url)
}
