% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mpan_details.R
\name{get_electric_mpan}
\alias{get_electric_mpan}
\title{Get mpan details
Get MPAN details from the output of the get_property_details function}
\usage{
get_electric_mpan(out, property_index = 1)
}
\arguments{
\item{out}{The output from the get_property_details function.}

\item{property_index}{The index of the property in the output list (default is 1).}
}
\value{
A list containing the MPAN, export MPAN, meter serial number, and export serial number.
}
\description{
Get mpan details
Get MPAN details from the output of the get_property_details function
}
\examples{
out <- get_property_details("your_account_number", "your_api_key")
}
