% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meter_details.R
\name{get_meter_info}
\alias{get_meter_info}
\title{Get Meter Details
This function retrieves details about a specific meter associated with an Octopus Energy account.}
\usage{
get_meter_info(acct_no, api_key)
}
\arguments{
\item{acct_no}{A string representing the account number.}

\item{api_key}{A string representing the API key for authentication.}
}
\value{
A list containing the meter details.
}
\description{
Get Meter Details
This function retrieves details about a specific meter associated with an Octopus Energy account.
}
\examples{
# get_meter_details("your_account_number", "your_api_key")
}
