% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tariff_details.R
\name{get_tariff_history}
\alias{get_tariff_history}
\title{Get Tariff Details
This function retrieves the tariff history for a specific property from the output of the get_property_details function.}
\usage{
get_tariff_history(out, property_index = 1)
}
\arguments{
\item{out}{The output from the get_property_details function.}

\item{property_index}{The index of the property in the output list (default is 1).}
}
\value{
A data frame containing the tariff history, including valid dates and duration.
}
\description{
Get Tariff Details
This function retrieves the tariff history for a specific property from the output of the get_property_details function.
}
\examples{
out <- get_meter_details("your_account_number", "your_api_key")
get_tariff_history(out, property_index = 1)
}
