% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tariffs.R
\name{get_tariff_info}
\alias{get_tariff_info}
\title{Get Tariff Information
This function retrieves tariff information associated with an Octopus Energy account.}
\usage{
get_tariff_info(api_key, acct, property_index)
}
\arguments{
\item{api_key}{A string representing the API key for authentication.}

\item{acct}{A string representing the account number.}

\item{property_index}{An integer representing the index of the property to retrieve tariff information for (default is 1).}
}
\value{
A list containing standing charges and unit rates for the current property consumption tariff.
}
\description{
Get Tariff Information
This function retrieves tariff information associated with an Octopus Energy account.
}
\examples{
get_tariff_info("your_account_number", "your_api_key", property_index = 1)
}
