% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helper.R
\name{octo_request}
\alias{octo_request}
\title{Construct a base Octopus API request}
\usage{
octo_request(url, api_key)
}
\arguments{
\item{url}{Full endpoint URL}

\item{api_key}{Octopus API key}
}
\value{
An httr2 request object
}
\description{
Adds authentication and a custom User-Agent header.
}
\keyword{internal}
